Timings Links
In case you may want to take a look at the timings paste as long as it doesn't get removed. Must've been something i've done wrongly causing player counts to be inaccurate. Anyways, the data is from my skyblock server and number of worlds is (mostly) proportional to the number of players.